Only check against the look vector
I think we can get away with only comparing the look vector of the
resultant basic rotation to that of the given matrix. We know that
that all three column vectors have a valid normal id, and we know that
the right and up vectors map to a valid rotation id. Only the look
vectors could possibly differ.