Write CFrame basic rotation IDs (#171)
* Add crate approx

* Add Vector3 method into_normal_id

* Fix identity Matrix3 constructor

* Add Matrix3 methods right_vector, up_vector, negative_look_vector

* Write CFrame basic rotation ids

Creates a module and moves CFrame-related logic to
it. Renames special_case_to_rotation to to_basic_rotation_id
Its counterpart, from_basic_rotation_id, goes in the other
direction from rotation matrices to ids.

Previously, the serializer would write out the entire rotation matrix
for basic rotations. It should now write the ids instead.

* Fix method name

* Forgot to stage...

* Only check against the look vector

I think we can get away with only comparing the look vector of the
resultant basic rotation to that of the given matrix. We know that
that all three column vectors have a valid normal id, and we know that
the right and up vectors map to a valid rotation id. Only the look
vectors could possibly differ.

* Add a comment

* Move look_vector_id binding back to where it was and reword slightly

We should exit immediately if the look vector does not map to a normal id.

* Don't bind rotation, reword again

* Rename negative_look_vector -> back_vector

* Remove TODO

We'd need to call this anyway, so it's fine to just compare the two
back vectors.

* Fix an off by one error of my own!

* Add Vector3 method from_normal_id

* Add Vector3 method cross

* Add Matrix3 method from_columns

* Strike _vector from these variable names

* No more match!

* Revert "No more match!"

This reverts commit 1f54deacb0e4badd48f04687dede2a04799dbf7e.

* Revert "Add Matrix3 method from_columns"

This reverts commit 4b385ccd24b90f3de0baaff30bb912a0752a81f6.

* Revert "Add Vector3 method cross"

This reverts commit 1a2d93c2fc8102bfd2b9f93c54e12a1c945319f4.

* Revert "Add Vector3 method from_normal_id"

This reverts commit a44520549d430093ed5e327f5fdf66ee3de2e962.

* Matrix3 fields are columns, not rows!

* Change the comment too

* Remove redundant absolute value

* Add a comment to Vector3::to_normal_id

* Wrap round-trip test in a module

* Use qualified path for cframe methods

* Remove crate approx

* Use copysign

To actually use this method here (which is not implemented for i32) we
have to first bind an f32 so that the compiler doesn't complain about
an ambiguous type, then cast to i32 when we're done.

* Move get_normal_id under to_normal_id

* Remove Matrix3 methods x_row, y_row, z_row, add method transpose

* I guess these need to be qualified?

Compiles on my machine...

* Use copysign, take 2