Fix support for empty Font tags in rbx_xml (#261)
In the past, Roblox serialized `Font` properties as empty tags. Although
they've stopped doing that, there are still models in the wild with that
format for tags. This adds support for empty Font tags by just returning
the default value for `Font` when the tag is empty.

Relies on https://github.com/rojo-rbx/rbx-test-files/pull/19.