Allow instance parents to be none Ref, use insert in WeakDom::new
Slight lifting of a restriction on WeawkDom::insert here - we'll
consider it OK to insert an instance with specifically *no* parent!
The only difference in behavior is that the inserted instance will not
be inserted into any parent's child list, but I think that's
reasonable and expectable

And all tests pass now, hooray!