Add more comprehensive test coverage to rbx_xml (#295)
During the rewrite for `rbx_xml`, I noticed that we have rather poor
test coverage for it. To resolve this, this PR adds more extensive test
coverage by using the same test files we do for `rbx_binary`.
Additionally, this adds a test for a 'file' with every data type that
checks formatting to ensure it never changes.