Implement IntoIterator for &Attributes and Extend for Attributes (#386)
The original implementation of Attributes had an `iter` method but it
did not implement `FromIterator` for `&Attributes`, making it so you
couldn't use `for (key, value) in &attributes` and had to use `for (key,
value) in attributes.iter()`. This changes that.

Additionally, `Extend` is a nice quality of life trait for containers so
it's implemented here.