Improve `rbx_reflector` Roblox Studio automation and logging on macOS (#407)
This PR brings automatic place saving to macOS, so it now works exactly the same
as on Windows. It also pipes the `stdout` and `stderr` streams to `null` to get
rid of extreme verbosity on macOS. Lastly, it adds a check to the injected
plugin so it won't run outside of the defaults place, which is useful when other
Roblox Studio instances are open and Roblox Studio's "Reload plugins on file
changed" setting is enabled.