Add command for stripping a specific property from Instances to rbx-util (#426)
When debugging an issue that cropped up on the Roblox devforums
yesterday, I wanted an easy way to remove every instance of property
from a file (specifically, `SurfaceAppearance.Color`). Given that
there's no easy way to do this without converting a file to XML,
removing the property, then converting it back, I think it'd be helpful
to include in rbx-util in the future.